[ATLAS-3367] Expose total count of index searches
Both ElasticSearch and SOLR can expose the total count
of results without returning all results. This is useful
to give the user or client an idea how many results there are.

This patch ensures the total is returned if available. This total
is an approximate as scrubbing of the results still needs to happen.
Therefore, one should not only rely on this information to provide
, for example, pagination.